refactor(component): rename menu buttons for clarity
