+++ packages/ui/menu/index.ts
@@ -1,4 +1,5 @@
-export * from './IconButton';
+export * from './IconMenuButton';
 export * from './List';
 export * from './Item';
 export * from './Menu';
+export * from './TextMenuButton';
